Update README.md
Start with a form and give it an ID attribute.   

Include the javaScript function at the top of the page 
 
Call the buildAuditChangesArray function in a document.ready function, using the form ID set in step #1

Ex.    
$(document).ready(function() {  
		$('#edit_form').buildAuditChangesArray(); 
	});
 
When the form is submitted, an array of form fields changes with an object containing the previous and new values is plugged into a hidden form field, "objAuditChange".
 
This has been tested extensively with different types of form inputs (ex. radio, text, textarea, checkbox, select).  With fields updated via javascript/ajax look into the trigger function (http://api.jquery.com/trigger/) and call focusin and change when fields are updated.